requirements.txt
requests==2.32.3
rich==13.8.1
jinja2==3.1.4
python-nmap==0.7.1
websocket-client==1.8.0
pyyaml==6.0.2
